testcases/llamaindexcross/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "agent"
authors = [{ name = "John Doe", email = "john.doe@myemail.com" }]
dependencies = [
    "uipath-llamaindex>=0.0.27",
    "uipath"
]
requires-python = ">=3.11"

[tool.uv.sources]
uipath = { path = "../../", editable = true }